rename CameraFollowTarget.js -> Followable.js, and split CameraFollowCoordinator out into its own file in the remote-operator-addon
