+++ index.html
@@ -168,7 +168,7 @@
     <script src="src/gui/ar/areaTargetScanner.js" type="module"></script>
     <script src="src/gui/ar/areaCreator.js" type="module"></script>
     <script src="src/gui/ar/videoPlayback.js" type="module"></script>
-    <script src="src/gui/ar/CameraFollowTarget.js" type="module"></script>
+    <script src="src/gui/ar/Followable.js" type="module"></script>
 
     <script src="src/gui/spatial/index.js" type="module"></script>
     <script src="src/gui/spatial/timeRecorder.js" type="module"></script>
